refactor(server): move four misfiled handlers into the files their routes imply
Motion only; no handler body changes.

internal/server/handlers_integrity.go was 78 lines holding exactly one of the
three integrity handlers (HandleGetDecisionProof, GET /v1/integrity/proof/{id}).
Its two siblings lived in handlers_decisions.go — HandleVerifyDecision
(GET /v1/verify/{id}) and HandleListIntegrityViolations
(GET /v1/integrity/violations) — even though all three routes are registered on
consecutive lines in server.go. The filename actively misdirected: a contributor
asked to change the violations response opens the file named "integrity", finds
the wrong handler, and has to grep.

HandleTraceHealth (GET /v1/trace-health) and HandleSessionView
(GET /v1/sessions/{session_id}) get their own files. Neither is a decision
resource, and both were sitting in the middle of the decisions file between
the integrity handlers and HandleRetractDecision.

The package convention is one file per resource and is otherwise well kept:
handlers_conflicts.go, handlers_runs.go, handlers_admin.go, handlers_keys.go,
handlers_assessments.go, handlers_retention.go, handlers_labels.go,
handlers_export.go, handlers_project_links.go, handlers_org_settings.go,
handlers_signup.go. These four were the exceptions.

The other sixteen handlers in handlers_decisions.go stay put. What remains is
one resource — trace, get, query, temporal query, agent history, search, check,
recent, revisions, retract, erase, timeline, facets, lineage, patch — with no
sub-entity to name a file after, and buildTraceAgentContext is already
correctly placed directly after its sole caller HandleTrace. 1,070 lines is not
by itself an argument for splitting further.

Verified as pure motion rather than asserted: the three moved blocks were
extracted before the edit and byte-compared against the destination files
afterwards, all three identical (105, 26 and 71 lines). server.go is untouched
— `git diff --stat` does not list it — because the route table refers to the
same method values on the same Handlers struct. `go test -list` over the
package is unchanged. Only the import blocks differ from pure motion, which
goimports settled: handlers_decisions.go drops strconv, integrity and
tracehealth, and the destinations gain what they use.

Build, vet and golangci-lint are clean. The package's one failing test,
TestGitCommitSubject_CurrentRepo, fails identically on main — it shells out to
git from the test process — and is unrelated to this change.
