+++ internal/server/handlers_decisions.go
@@ -5,7 +5,6 @@ import (
 	"errors"
 	"fmt"
 	"net/http"
-	"strconv"
 	"strings"
 	"sync"
 	"time"
@@ -15,11 +14,9 @@ import (
 	"github.com/ashita-ai/akashi/internal/auth"
 	"github.com/ashita-ai/akashi/internal/authz"
 	"github.com/ashita-ai/akashi/internal/compact"
-	"github.com/ashita-ai/akashi/internal/integrity"
 	"github.com/ashita-ai/akashi/internal/model"
 	"github.com/ashita-ai/akashi/internal/service/decisions"
 	"github.com/ashita-ai/akashi/internal/service/quality"
-	"github.com/ashita-ai/akashi/internal/service/tracehealth"
 	"github.com/ashita-ai/akashi/internal/storage"
 )
 
@@ -745,211 +742,6 @@ func (h *Handlers) HandleDecisionRevisions(w http.ResponseWriter, r *http.Reques
 		DecisionID: id,
 		Revisions:  revisions,
 		Count:      len(revisions),
-	})
-}
-
-// HandleVerifyDecision handles GET /v1/verify/{id}.
-// Recomputes the SHA-256 content hash from stored fields and compares to the stored hash.
-func (h *Handlers) HandleVerifyDecision(w http.ResponseWriter, r *http.Request) {
-	orgID := OrgIDFromContext(r.Context())
-
-	id, err := parsePathUUID(r, "id")
-	if err != nil {
-		writeError(w, r, http.StatusBadRequest, model.ErrCodeInvalidInput, "invalid decision ID")
-		return
-	}
-
-	claims := ClaimsFromContext(r.Context())
-
-	d, err := h.db.GetDecision(r.Context(), orgID, id, storage.GetDecisionOpts{})
-	if err != nil {
-		writeError(w, r, http.StatusNotFound, model.ErrCodeNotFound, "decision not found")
-		return
-	}
-
-	ok, err := canAccessAgent(r.Context(), h.db, claims, d.AgentID)
-	if err != nil {
-		h.writeInternalError(w, r, "authorization check failed", err)
-		return
-	}
-	if !ok {
-		writeError(w, r, http.StatusForbidden, model.ErrCodeForbidden, "no access to this decision")
-		return
-	}
-
-	resp := model.VerifyDecisionResponse{DecisionID: id}
-
-	switch {
-	case d.ValidTo != nil:
-		resp.Status = "retracted"
-		resp.RetractedAt = d.ValidTo.UTC().Format(time.RFC3339Nano)
-		if d.ContentHash == "" {
-			verified := false
-			resp.Verified = &verified
-			resp.Message = "this decision was created before content hashing was enabled"
-		} else {
-			valid := integrity.VerifyContentHash(d.ContentHash, d.ID, d.DecisionType, d.Outcome, d.Confidence, d.Reasoning, d.ValidFrom)
-			resp.Verified = &valid
-			resp.ContentHash = d.ContentHash
-		}
-	case d.ContentHash == "":
-		resp.Status = "no_hash"
-		resp.Message = "this decision was created before content hashing was enabled"
-	default:
-		erasure, erasureErr := h.db.GetDecisionErasure(r.Context(), orgID, id)
-		switch {
-		case erasureErr == nil:
-			valid := integrity.VerifyContentHash(d.ContentHash, d.ID, d.DecisionType, d.Outcome, d.Confidence, d.Reasoning, d.ValidFrom)
-			resp.Status = "erased"
-			resp.Valid = &valid
-			resp.ContentHash = d.ContentHash
-			resp.OriginalHash = erasure.OriginalHash
-			resp.ErasedAt = &erasure.ErasedAt
-			resp.ErasedBy = erasure.ErasedBy
-		case !isNotFoundError(erasureErr):
-			h.writeInternalError(w, r, "failed to check erasure status", erasureErr)
-			return
-		default:
-			valid := integrity.VerifyContentHash(d.ContentHash, d.ID, d.DecisionType, d.Outcome, d.Confidence, d.Reasoning, d.ValidFrom)
-			resp.Valid = &valid
-			if valid {
-				resp.Status = "verified"
-			} else {
-				resp.Status = "tampered"
-			}
-			resp.ContentHash = d.ContentHash
-		}
-	}
-
-	writeJSON(w, r, http.StatusOK, resp)
-}
-
-// HandleListIntegrityViolations handles GET /v1/integrity/violations.
-// Returns recent integrity violations for the caller's organization, ordered
-// newest-first. This exposes the durable audit trail written by the background
-// integrity audit loop (auditIntegrityProofs). Admin-only because violations
-// indicate potential tampering and are part of incident response.
-func (h *Handlers) HandleListIntegrityViolations(w http.ResponseWriter, r *http.Request) {
-	orgID := OrgIDFromContext(r.Context())
-
-	limit := 50
-	if v := r.URL.Query().Get("limit"); v != "" {
-		n, err := strconv.Atoi(v)
-		if err != nil || n < 1 || n > 200 {
-			writeError(w, r, http.StatusBadRequest, model.ErrCodeInvalidInput, "limit must be between 1 and 200")
-			return
-		}
-		limit = n
-	}
-
-	violations, err := h.db.GetIntegrityViolations(r.Context(), orgID, limit)
-	if err != nil {
-		h.writeInternalError(w, r, "failed to list integrity violations", err)
-		return
-	}
-
-	writeJSON(w, r, http.StatusOK, model.IntegrityViolationsResponse{
-		Violations: violations,
-		Count:      len(violations),
-	})
-}
-
-// HandleTraceHealth handles GET /v1/trace-health.
-// Returns aggregate health metrics for the caller's organization.
-// Optional query params: from, to (RFC3339) scope metrics to a time window.
-func (h *Handlers) HandleTraceHealth(w http.ResponseWriter, r *http.Request) {
-	orgID := OrgIDFromContext(r.Context())
-
-	from, err := queryTime(r, "from")
-	if err != nil {
-		writeError(w, r, http.StatusBadRequest, model.ErrCodeInvalidInput, err.Error())
-		return
-	}
-	to, err := queryTime(r, "to")
-	if err != nil {
-		writeError(w, r, http.StatusBadRequest, model.ErrCodeInvalidInput, err.Error())
-		return
-	}
-
-	svc := tracehealth.New(h.db)
-	metrics, err := svc.Compute(r.Context(), orgID, from, to)
-	if err != nil {
-		h.writeInternalError(w, r, "failed to compute trace health", err)
-		return
-	}
-
-	writeJSON(w, r, http.StatusOK, metrics)
-}
-
-// HandleSessionView handles GET /v1/sessions/{session_id}.
-// Returns all decisions from a given MCP/HTTP session, with summary statistics.
-func (h *Handlers) HandleSessionView(w http.ResponseWriter, r *http.Request) {
-	claims := ClaimsFromContext(r.Context())
-	orgID := OrgIDFromContext(r.Context())
-
-	sid, err := parsePathUUID(r, "session_id")
-	if err != nil {
-		writeError(w, r, http.StatusBadRequest, model.ErrCodeInvalidInput, "invalid session_id")
-		return
-	}
-
-	decs, err := h.db.GetSessionDecisions(r.Context(), orgID, sid)
-	if err != nil {
-		h.writeInternalError(w, r, "failed to get session decisions", err)
-		return
-	}
-
-	decs, err = filterDecisionsByAccess(r.Context(), h.db, claims, decs, h.grantCache)
-	if err != nil {
-		h.writeInternalError(w, r, "authorization check failed", err)
-		return
-	}
-
-	if len(decs) == 0 {
-		writeJSON(w, r, http.StatusOK, model.SessionViewResponse{
-			SessionID:     sid,
-			Decisions:     []model.Decision{},
-			DecisionCount: 0,
-		})
-		return
-	}
-
-	// Compute summary: use min/max of valid_from to avoid ordering edge cases
-	// (multiple decisions can share the same valid_from in revision chains).
-	startedAt := decs[0].ValidFrom
-	endedAt := decs[0].ValidFrom
-	for _, d := range decs[1:] {
-		if d.ValidFrom.Before(startedAt) {
-			startedAt = d.ValidFrom
-		}
-		if d.ValidFrom.After(endedAt) {
-			endedAt = d.ValidFrom
-		}
-	}
-	duration := endedAt.Sub(startedAt).Seconds()
-	if duration < 0 {
-		duration = 0
-	}
-
-	decisionTypes := map[string]int{}
-	var totalConf float64
-	for _, d := range decs {
-		decisionTypes[d.DecisionType]++
-		totalConf += float64(d.Confidence)
-	}
-	avgConfidence := totalConf / float64(len(decs))
-
-	writeJSON(w, r, http.StatusOK, model.SessionViewResponse{
-		SessionID:     sid,
-		Decisions:     decs,
-		DecisionCount: len(decs),
-		Summary: &model.SessionViewSummary{
-			StartedAt:     startedAt,
-			EndedAt:       endedAt,
-			DurationSecs:  duration,
-			DecisionTypes: decisionTypes,
-			AvgConfidence: avgConfidence,
-		},
 	})
 }
